requirements.txt
customtkinter
pyperclip
pygame
cryptography
qrcode
Pillow